Ensure "Take the Test" button is always enabled
Previously, this button was disabled by default and then only enabled on
the first page load (not when navigating home through the site).  Even
on that load it was only enabled after an XHR call to an endpoint on our
server to retrieve location information.

This commit fixes this with the following changes:
* The button is enabled by default and still works when navigating back.
* The location information is retrieved when the button is clicked
  rather than immediately on load.