Remove "Take the Test" link from the header
This link didn't function previously.  It would have been fairly simple
to fix it on the home page, but seems that it would take a larger rework
for it to work when clicked from other pages.  For now, we're removing
this broken functionality, though we'll later want to point users toward
the test from every page.