Start location check as soon as homepage loads
We need to call `navigator.geolocation.getCurrentPosition`, pass it's
data to our /home/get_location_data endpoint, and receive a response
before the speed test can be run.  This process can take surprisingly
long and the user is blocked from running the test until it has
completed. To improve the user experience, we're staring the process as
soon as the home page loads instead of waiting for the user to click the
"Take the Test!!" button first.