Don't set coord fields from ndt_speed_test.coffee
The form fields with ids submission_latitude and submission_longitude
are already set from set_coords() in home.coffee, which is also
responsible for enabling the button that starts the speed test.  We
don't need to duplicate this functionality; the call chain for both of
these async functions begin at the same point in home.coffee.