Update messaging for failed geolocation lookup
Firefox's geolocation hasn't been working, but it's not actually an
issue with the browser itself. The issue is that the Google API it
depends on to look up the geolocation is denying the request. There
appears to be a repeating issue over the years of API keys for that
service hitting service limits and fixes have generally been slow.

A couple bugzilla links related to this issue, if you're curious:
https://bugzilla.mozilla.org/show_bug.cgi?id=1232995
https://bugzilla.mozilla.org/show_bug.cgi?id=1610306

Because this isn't something we can fix for our users, the best we can
offer is the option to change their browser configuration to use a
different location service. Mozilla's location service seems to be a
little less accurate, but it's still fairly close. In my test, it showed
an area that was in my current city. For comparison, an IP address
lookup I did returned an entirely different city an hour's drive away.