Enable starting the test via the relevant nav link
By checking the URL hash, we're now able to start the test right away,
even if the "Take the Test" button in the nav menu was clicked while on
a page other than the home page. Before this commit, that link would
only scroll down to the main "Take the Test!!" button on the home page.